+++ src/utils/config.py
@@ -13,7 +13,7 @@ class ConfigWallet(TypedDict):
 
     max_gas_limit: int
     private_key: str
-    allowed_errors: Optional[list[str]]
+    allowed_sim_errors: Optional[list[str]]
 
 
 class ConfigSnapshotSync(TypedDict):